+++ chambre.h
@@ -8,30 +8,30 @@ class Chambre
 {
    private:
        int idchambre;
-       int tarif;
+       float tarif;
        QString etat,type;
 public:
 
     Chambre();
-    Chambre(int idchambre,QString etat,QString type,int tarif );
+    Chambre(int idchambre,QString etat,QString type,float tarif );
 
 
     QString getetat(){return etat ;}
     QString gettype(){return type ;}
     int getidchambre(){return idchambre;}
-    int gettarif(){return tarif;}
+    float gettarif(){return tarif;}
 
 
     void setetat(QString e){etat =e;}
     void settype(QString ty){type = ty;}
     void setidchambre (int idchambre ){this->idchambre=idchambre;}
-    void settarif (int tarif ){this->tarif=tarif;}
+    void settarif (float tarif ){this->tarif=tarif;}
 
         // fonctionnalites de base relatives a l'entite chambre
     bool ajouter ();
     bool supprimer(int idchambre);
     QSqlQueryModel * afficher();
-    bool modifier(int idchambre, QString etat, QString type, int tarif);
+    bool modifier(int idchambre, QString etat, QString type, float  tarif);
